% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSEERParser.R
\name{buildSEERParser}
\alias{buildSEERParser}
\title{Builds a parser to analyze SEER data.}
\usage{
buildSEERParser(file_path, file_source)
}
\arguments{
\item{file_path}{A path to the .sas file or .dic file.}

\item{file_source}{If it is a dictionary file from SEER*Stat the option is
'seerstat'. Otherwise the option is 'download'.}
}
\value{
A list with parsing instructions for \link{readSEER}.
}
\description{
Datasets downloaded from SEER website (after the request process) come in
a compressed file containing, among other things, a .sas file with
instructions on data fields (fields width) to parse the text files and also
one sentence explanation for every field. One can also obtain data from SEER
through the SEER*Stat software. In this case, the text file may come in
different ways that are explained in a dictionary file (.dic).
}
\examples{
\dontrun{
instr <- buildSEERParser(file_path = "read.seer.research.nov17.sas",
                         file_source = 'download')}

}
