% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSingleValueFeatures.R
\name{findSingleValueFeatures}
\alias{findSingleValueFeatures}
\title{Find features with a single valeu for all rows}
\usage{
findSingleValueFeatures(dataframe)
}
\arguments{
\item{dataframe}{The dataframe with SEER data}
}
\description{
Some features may have the same value for all rows depending on the way you
perform a subset. In these cases, this feature is not adding information to
the analysis and therefore should be removed. This function helps you to
find such features and provides you also the unique value of such a feature.
}
\details{
You can provide the return$`Unique value` of this function as the argument
to the \code{removeFullNAFeatures} function for even if an unique value for
a feature is also a value in another feature, it will only be removed if it
is the only valeu for all rows for this specific feature.
}
