% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNormalizedEntropy.R
\name{getNormalizedEntropy}
\alias{getNormalizedEntropy}
\title{Gets the normalized entropy for features in a data frame}
\usage{
getNormalizedEntropy(dataframe)
}
\arguments{
\item{dataframe}{The dataframe with SEER data.}
}
\description{
Gets the normalized entropy for every feature in a data frame.
}
