% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSEER.R
\name{listPrimarySites}
\alias{listPrimarySites}
\title{Lists cancers primary sites that are supported by this package}
\usage{
listPrimarySites()
}
\value{
It displays on the screen the cancer primary sites supported by this
  package.
}
\description{
Lists cancers primary sites that are supported by this package
}
\seealso{
\url{https://seer.cancer.gov/siterecode/icdo3_dwhoheme/index.html}
}
