% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessSEER.R
\name{preprocessSEER}
\alias{preprocessSEER}
\title{Reads a sas instruction file in order to parse SEER ASCII data files}
\usage{
preprocessSEER(file_path, seerstats = FALSE)
}
\arguments{
\item{file_path}{A path to the SAS file}
}
\value{
A data frame with instruction for readSEER
}
\description{
Reads a sas instruction file in order to parse SEER ASCII data files
}
\examples{
# You must preprocess a SAS file before reading the data with readSEER
# Do not run
# preprocessSEER("read.seer.research.nov17.sas")
}
