% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeFullNAFeatures.R
\name{removeFullNAFeatures}
\alias{removeFullNAFeatures}
\title{Remove full-NA value features in a SEER dataframe}
\usage{
removeFullNAFeatures(dataframe, additional_na)
}
\arguments{
\item{dataframe}{The dataframe with SEER data}

\item{additional_na}{A vector with additional symbol(s) that should also be
considered NA. This is important for some datasets exported from SEER*Stat
software that come with NA values and also strings 'Blank(s)' representing
also lack of values.}
}
\description{
Some features in a SEER dataframe are specific to certain types of cancer.
If you're analyzing only patients with breast cancer, features that have
information on prostate cancer, for example, will have only NA values. These
features are useless to your analysis and should be removed. \code{histNA}
is a good function to plot this scenario so that you know if you have
features with all their rows as NA values. By making use of
removeFullNAFeatures you can get rid of these features.
}
