% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSEER.R
\name{siteLookUp}
\alias{siteLookUp}
\title{Converts primary site name to recode}
\usage{
siteLookUp(site_name)
}
\arguments{
\item{site_name}{cancer primary site name}
}
\value{
Recode code for the specified cancer primary site name
}
\description{
Converts primary site name to recode
}
