% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSEER.R
\name{listPrimarySites}
\alias{listPrimarySites}
\title{List cancers primary sites that are supported by this package}
\usage{
listPrimarySites()
}
\value{
It displays on the screen the cancer primary sites supported by the package
}
\description{
List cancers primary sites that are supported by this package
}
