% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessSEER.R
\name{preprocessSEER}
\alias{preprocessSEER}
\title{Preprocess SEER instruction files.}
\usage{
preprocessSEER(file_path, filetype)
}
\arguments{
\item{file_path}{A path to the .sas file or .dic file from SEER*Stat.}

\item{filetype}{The type of instruction file you are providing. If it is a
a dictionary file from SEER*Stat the option is 'dictionary'. Otherwise the
option is 'sas'.}
}
\value{
A list with instructions for \link{readSEER}.
}
\description{
Datasets downloaded from SEER website (after the request process) come in
a compressed file containing, among other things, a .sas file with
instructions on data fields (fields width) to parse the text files and also
one sentence explanation for every field. One can also obtain data from SEER
through the SEER*Stat software. In this case, the text file may come in
different ways in a dictionary file (.dic).
}
\examples{
\dontrun{
instr <- preprocessSEER("read.seer.research.nov17.sas", filetype = 'sas')}

}
