% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSEER.R
\name{readSEER}
\alias{readSEER}
\title{Reads SEER data from ASCII data file(s).}
\usage{
readSEER(path, instructions, read_dir = FALSE, year_dx, primary_site)
}
\arguments{
\item{path}{A path or a vector with paths to either folder with SEER ASCII
data files or directly to the file(s).}

\item{instructions}{The output variable from the buildSEERParser function}

\item{read_dir}{If true, it reads all .TXT files in the directory or
directories provided in the path parameter. If this parameter is missing
or FALSE, vidente will only read the TXT you explicitly informed. This
function does not work with data exported from the SEER*Stat software.}

\item{year_dx}{Filter the result to this year of diagnosis (or range of
years of diagnosis)}

\item{primary_site}{Filter the result to a primary site (or a vector of
primary sites for SEER*Stat files, or a site category from
listPrimarySite() for data downloaded from SEER).}
}
\value{
A data frame with all SEER data from the ASCII data file(s) you
  provided, given the criteria you chose in the parameters.
}
\description{
Reads SEER data from ASCII data file(s).
}
\examples{
# First build parsing instructions
\dontrun{
instr <- buildSEERParser(file_path = 'read.seer.research.nov17.sas',
                         file_source = 'download')

# Now you can read it
paths <- c('/home/yourusername/SEER/yr1973_2015.seer9/BREAST.TXT',
          '/home/yourusername/SEER/yr2000_2015.ca_ky_lo_nj_ga/BREAST.TXT')

# I'm interested here in patients with breast cancer diagnosed between 2012
# and 2015
seer_data <- readSEER(path = paths,
                      instructions = instr,
                      year_dx = c(2012:2015),
                      primary_site = 'Breast')}

}
