% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vidente.R
\docType{package}
\name{vidente}
\alias{vidente}
\alias{vidente-package}
\title{vidente: A package for parsing and analyzing SEER data.}
\description{
The vidente package provides two categories of important functions.
Functions to parse SEER data and functions to analyze it.
}
\section{Parsing data}{

The foo functions ...
}

\section{Analyzing data}{

The bar functions ...
}

